# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/email/{email}": {
      "get": {
        "description": "Fetches a single user using their email address.",
        "operationId": "UserController_getUserByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "description": "Email address of the user",
            "schema": {
              "example": "example@email.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Get user by email",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/crn/{crn}": {
      "get": {
        "description": "Fetches a user using their unique CRN.",
        "operationId": "UserController_getUserByCRN",
        "parameters": [
          {
            "name": "crn",
            "required": true,
            "in": "path",
            "description": "Customer Registration Number of the user",
            "schema": {
              "example": "0123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Get user by their CRN",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/name": {
      "get": {
        "description": "Returns a list of users that match the provided name.",
        "operationId": "UserController_getUserByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "description": "Name (or partial name) to search users by",
            "schema": {
              "example": "Sarah",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Name parameter is required",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No users found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "No users found with the name",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Search users by name",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "description": "Returns the currently authenticated user's data.",
        "operationId": "UserController_getCurrentUserData",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "JWT token must be set in the cookie named \"token\". Example: token=your_jwt_token_here",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current user data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: Token missing or invalid/expired.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "No token found in cookies",
                        "error": "Unauthorized"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "Invalid or expired token",
                        "error": "Unauthorized"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "cookie": []
          },
          {
            "bearer": []
          },
          {
            "token": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Updates the currently authenticated user's profile information.",
        "operationId": "UserController_updateCurrentUserData",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "JWT token must be set in the cookie named \"token\". Example: token=your_jwt_token_here",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "User data to update (you don't have to not send all fields, only the ones you want to change will be updated).",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "These categories are invalid: Hi, LOL",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: Token missing or invalid/expired.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "No token found in cookies",
                        "error": "Unauthorized"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "Invalid or expired token",
                        "error": "Unauthorized"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "cookie": []
          },
          {
            "bearer": []
          },
          {
            "token": []
          }
        ],
        "summary": "Update current user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/profile-picture": {
      "post": {
        "description": "Uploads a new profile picture for the authenticated user and updates their record. Replaces the previous picture if it exists.<br><strong>Note:</strong> The file must be an image (PNG, JPEG, WebP) and cannot exceed 5MB in size.<br><strong>IMPORTANT:</strong> Please ensure that uploaded images comply with Islamic laws. This means avoiding haram content such as music-related images, depictions of women's bodies (even hands), or any illustrations of living beings (humans, animals, etc.) whether drawn or digital.",
        "operationId": "UserController_updateProfilePicture",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "JWT token must be set in the cookie named \"token\". Example: token=your_jwt_token_here",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Profile picture file to upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile picture updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Profile picture updated successfully",
                    "pfpFileName": "user-avatar/12345.png"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. File missing or invalid format.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "No file provided or invalid file type",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: Token missing or invalid/expired.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "No token found in cookies",
                        "error": "Unauthorized"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "Invalid or expired token",
                        "error": "Unauthorized"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found or profile picture cannot be updated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "User not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "cookie": []
          },
          {
            "bearer": []
          },
          {
            "token": []
          }
        ],
        "summary": "Update user profile picture",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Deletes the authenticated user's current profile picture and replaces it with a default avatar.",
        "operationId": "UserController_deleteProfilePicture",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "description": "JWT token must be set in the cookie named \"token\". Example: token=your_jwt_token_here",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile picture deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Profile picture deleted successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: Token missing or invalid/expired.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "No token found in cookies",
                        "error": "Unauthorized"
                      }
                    },
                    {
                      "example": {
                        "statusCode": 401,
                        "message": "Invalid or expired token",
                        "error": "Unauthorized"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found or already has the default profile picture",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Profile picture not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "cookie": []
          },
          {
            "bearer": []
          },
          {
            "token": []
          }
        ],
        "summary": "Delete current user profile picture",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/profile-picture": {
      "get": {
        "description": "Retrieves the URL of a user's profile picture that is stored on R2. Throws a 404 error if the user or their profile picture does not exist.",
        "operationId": "UserController_getUserProfilePictureUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the user whose profile picture URL is being retrieved",
            "schema": {
              "example": "b4f2d8c1-9876-5432-10ba-2f3e4d5c6789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile picture URL retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "pfpUrl": "https://cdn.example.com/pfp/USER_ID.jpeg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20250101T000000Z&X-Amz-Expires=3600&X-Amz-Signature=REDACTED"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User or profile picture not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Profile picture not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Get user's profile picture URL",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/profile-pictures/batch": {
      "post": {
        "description": "Retrieves the profile picture URLs for multiple users by their IDs. Only valid UUIDs are considered.",
        "operationId": "UserController_getUsersProfilePicturesUrls",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Array of user IDs to retrieve profile picture URLs for",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "example": [
                      "a3e1c9f0-1234-5678-90ab-1c2d3e4f5678",
                      "b4f2d8c1-9876-5432-10ba-2f3e4d5c6789"
                    ]
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile picture URLs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "a3e1c9f0-1234-5678-90ab-1c2d3e4f5678",
                      "pfpUrl": "https://cdn.example.com/pfp/USER_ID.jpeg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20250101T000000Z&X-Amz-Expires=3600&X-Amz-Signature=REDACTED"
                    },
                    {
                      "id": "b4f2d8c1-9876-5432-10ba-2f3e4d5c6789",
                      "pfpUrl": null
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "No user IDs provided",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No users found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "No users found with the provided IDs",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "summary": "Get profile pictures for multiple users",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Silah Backend API Documentation",
    "description": "Use this documentation to explore, test, and understand the available API endpoints, their request/response structure, and any required parameters such as headers, cookies, or authentication tokens.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clv70z13w0000unqoj4lcr8x4"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "crn": {
            "type": "string",
            "example": "1234567890"
          },
          "businessName": {
            "type": "string",
            "example": "Acme Corp"
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "SUPPLIER",
              "BUYER"
            ],
            "example": "BUYER"
          },
          "city": {
            "type": "string",
            "example": "Riyadh"
          },
          "pfpFileName": {
            "type": "string",
            "example": "moon-3263ec71-3e92-441d-aadd-a57b4a99b2e2.jpeg",
            "description": "Profile picture file name in R2 bucket."
          },
          "pfpUrl": {
            "type": "string",
            "example": "https://gp-silah.d025be9440ae5eb8295c69a36497276a.r2.cloudflarestorage.com/gp-silah/moon.jpeg-30510246-41f7-4cff-a052-78bcc30f7301.jpeg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Content-Sha256=UNSIGNED-PAYLOAD&X-Amz-Credential=...&X-Amz-Date=20250816T131236Z&X-Amz-Expires=3600&X-Amz-Signature=...",
            "description": "Example of a signed URL from R2. Signed URLs expire 1 hour after creation.",
            "format": "uri"
          },
          "categories": {
            "example": [
              "Home & Living",
              "Technical & Repair Services"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-04T7:31:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-07-05T14:48:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "crn",
          "businessName",
          "role",
          "city",
          "pfpFileName",
          "pfpUrl",
          "categories",
          "isEmailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "maxLength": 25
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "newPassword": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8,
            "maxLength": 28
          },
          "businessName": {
            "type": "string",
            "example": "Acme Corp",
            "maxLength": 50
          },
          "city": {
            "type": "string",
            "example": "Riyadh"
          },
          "categories": {
            "example": [
              "Home & Living",
              "Technical & Repair Services"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}